move Address to core module
