+++ grpc/src/client/load_balancing/round_robin.rs
@@ -282,9 +282,9 @@ mod test {
     use crate::client::load_balancing::test_utils::TestEvent;
     use crate::client::load_balancing::test_utils::TestWorkScheduler;
     use crate::client::load_balancing::test_utils::{self};
-    use crate::client::name_resolution::Address;
     use crate::client::name_resolution::Endpoint;
     use crate::client::name_resolution::ResolverUpdate;
+    use crate::core::Address;
     use crate::metadata::MetadataMap;
     use crate::rt::default_runtime;
