+++ grpc/src/inmemory/mod.rs
@@ -49,7 +49,6 @@ use crate::client::ResponseStreamItem;
 use crate::client::SendOptions as ClientSendOptions;
 use crate::client::SendStream as ClientSendStream;
 use crate::client::Trailers as ClientTrailers;
-use crate::client::name_resolution::Address;
 use crate::client::name_resolution::ChannelController as ResolverChannelController;
 use crate::client::name_resolution::Endpoint;
 use crate::client::name_resolution::Resolver;
@@ -63,6 +62,7 @@ use crate::client::transport::GLOBAL_TRANSPORT_REGISTRY;
 use crate::client::transport::SecurityOpts;
 use crate::client::transport::Transport;
 use crate::client::transport::TransportOptions;
+use crate::core::Address;
 use crate::core::RecvMessage;
 use crate::core::SendMessage;
 use crate::credentials::SecurityLevel;
